updated add-location route
removed most of the get route because this would only be used by the back end to display the zipcode entry form for testing only.  User auth can be checked in the post route.  The post route is working, adding a zipcode, userid, lat, long.  is not secured as of now.